+++ programs/drift/src/controller/pnl.rs
@@ -26,6 +26,7 @@ use crate::state::margin_calculation::MarginContext;
 
 use crate::state::events::{OrderActionExplanation, SettlePnlExplanation, SettlePnlRecord};
 use crate::state::oracle_map::OracleMap;
+use crate::state::paused_operations::PerpOperation;
 use crate::state::perp_market::MarketStatus;
 use crate::state::perp_market_map::PerpMarketMap;
 use crate::state::spot_market::{SpotBalance, SpotBalanceType};
@@ -139,6 +140,20 @@ pub fn settle_pnl(
         ErrorCode::InvalidMarketStatusToSettlePnl,
         "Cannot settle pnl under current market status"
     )?;
+
+    validate!(
+        !perp_market.is_operation_paused(PerpOperation::SettlePnl),
+        ErrorCode::InvalidMarketStatusToSettlePnl,
+        "Cannot settle pnl under current market status"
+    )?;
+
+    if user.perp_positions[position_index].base_asset_amount != 0 {
+        validate!(
+            !perp_market.is_operation_paused(PerpOperation::SettlePnlWithPosition),
+            ErrorCode::InvalidMarketStatusToSettlePnl,
+            "Cannot settle pnl with position under current market status"
+        )?;
+    }
 
     let pnl_pool_token_amount = get_token_amount(
         perp_market.pnl_pool.scaled_balance,